pyproject.toml
[tool.poetry]
name = "newsaggregatorservice"
version = "0.1.0"
description = "Deliver news to Atomic FE"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
